IoTCloudServices/microservices/routes_microservice/code/requirements.txt
requests
flask
flask_cors
mysql-connector-python